Rename src/services/cellml.js to src/services/compress.js.
